Kept velruse.api, updated tests.
