+++ tests/selenium/testing.ini
@@ -1,7 +1,7 @@
 [testconfig]
 selenium.driver = firefox
 
-base_url = http://localhost:5000
+login_url = http://localhost:5000/login
 
 test_providers =
 #    facebook
@@ -30,9 +30,11 @@ test_providers =
 
 #google.login =
 #google.password =
+#google.login_url =
 
 #yahoo.login =
 #yahoo.password =
+#yahoo.login_url =
 
 #live.app =
 #live.login =
@@ -41,52 +43,49 @@ test_providers =
 [app:main]
 use = egg:testapp
 
-providers =
+login_providers =
 #    facebook
 
-# OpenID storage required by:
-# google, yahoo and openid providers
-#velruse.openid.store = openid.store.memstore.MemoryStore
-#velruse.openid.realm =
-
 # Bitbucket
-#velruse.bitbucket.consumer_key =
-#velruse.bitbucket.consumer_secret =
-#velruse.bitbucket.authorize = true
+#bitbucket.consumer_key =
+#bitbucket.consumer_secret =
+#bitbucket.authorize = true
 
 # Facebook
-#velruse.facebook.app_id =
-#velruse.facebook.app_secret =
-#velruse.facebook.scope =
+#facebook.app_id =
+#facebook.app_secret =
+#facebook.scope =
 
 # Github
-#velruse.github.app_id =
-#velruse.github.app_secret =
-#velruse.github.scope =
+#github.app_id =
+#github.app_secret =
+#github.scope =
 
 # Google (also requires OpenID configuration)
-#velruse.google.consumer_key =
-#velruse.google.consumer_secret =
-#velruse.google.oauth_scope =
+#google.consumer_key =
+#google.consumer_secret =
+#google.oauth_scope =
+#google.realm =
 
 # LinkedIn
-#velruse.linkedin.consumer_key =
-#velruse.linkedin.consumer_secret =
-#velruse.linkedin.authorize =
+#linkedin.consumer_key =
+#linkedin.consumer_secret =
+#linkedin.authorize =
 
 # Windows Live
-#velruse.live.client_id =
-#velruse.live.client_secret =
-#velruse.live.scope =
+#live.client_id =
+#live.client_secret =
+#live.scope =
 
 # Twitter
-#velruse.twitter.consumer_key =
-#velruse.twitter.consumer_secret =
-#velruse.twitter.authorize =
+#twitter.consumer_key =
+#twitter.consumer_secret =
+#twitter.authorize =
 
 # Yahoo (also requires OpenID configuration)
-#velruse.yahoo.consumer_key =
-#velruse.yahoo.consumer_secret =
+#yahoo.consumer_key =
+#yahoo.consumer_secret =
+#yahoo.realm =
 
 [server:main]
 use = egg:Paste#http
